Change package of GlazedListTableModel
